# Refactor protocol implementation by using crate bytes and fastrand