--- a/Cargo.toml
+++ b/Cargo.toml
@@ -17,10 +17,11 @@ nightly = []
 [dependencies]
 byteorder = "1.2"
 semver = "0.9"
-rand = "0.7"
+fastrand = "1.3"
 conhash = "0.4"
 log = "0.4"
 bufstream = "0.1"
+bytes = "0.5"
 
 [target.'cfg(unix)'.dependencies]
 unix_socket = "0.5"