--- a/src/client/mod.rs
+++ b/src/client/mod.rs
@@ -14,6 +14,7 @@ use std::collections::{BTreeMap, HashMap};
 use std::io;
 use std::net::TcpStream;
 use std::ops::Deref;
+use std::path::Path;
 use std::rc::Rc;
 
 use conhash::{ConsistentHash, Node};