… src/proto/binarydef.rs