Replace `rand` with `fastrand` which is smaller and faster.

Replace `Vec<u8>` in `RequestPacket` and `ResponsePacket` with `Bytes` in crate [bytes](https://github.com/tokio-rs/bytes).
> Bytes is an efficient container for storing and operating on contiguous slices of memory.

By using it, we can allocate a block of memory then split it to 3 slices instead of 3 individual allocation.
The drawback is `Bytes` can't grow up in place. And the current public APIs isn't changed yet, there's still `Vec<u8>` creation which is inefficient. I think it's worth to return `Bytes` directly to users, if they want to resize it they could turn it into  `Vec<u8>`.