@songzhi 2020-08-26
I'll do some benchmarks.