@songzhi 2020-08-26
It seems do have performance improvements. I didn't test memory usage, but in theory at least the new implementation wouldn't be worse than the old one.
Here's my benchmark code.
```rust
use criterion::Criterion;
use bytes::Bytes;
use std::io::Read;

fn request_packet(size: usize) -> Vec<u8> {
    let mut buf = vec![];
    let key = Bytes::from(vec![0u8; size]);
    let value = Bytes::from(vec![0u8; size * 3]);
    let req_packet = RequestPacket::new(
        Command::Set,
        DataType::RawBytes,
        0,
        0,
        0,
        vec![0xde, 0xad, 0xbe, 0xef, 0x00, 0x00, 0x0e, 0x1
        key,
        value,
    );
    req_packet.write_to(&mut buf).unwrap();
    buf
}
pub fn benches(c: &mut Criterion) {
    let small_packet = request_packet(32);
    let medium_packet = request_packet(128);
    let large_packet = request_packet(1024);
    c.bench_function("new:small", |b| {
        b.iter(|| {
            let mut reader = small_packet.as_slice();
            RequestPacket::read_from(&mut reader).unwrap()
        });
    });
    c.bench_function("new:medium", |b| {
        b.iter(|| {
            let mut reader = medium_packet.as_slice();
            RequestPacket::read_from(&mut reader).unwrap()
        });
    });
    c.bench_function("new:large", |b| {
        b.iter(|| {
            let mut reader = large_packet.as_slice();
            RequestPacket::read_from(&mut reader).unwrap()
        });
    });
    c.bench_function("old:small", |b| {
        b.iter(|| {
            let mut reader = small_packet.as_slice();
            OldRequestPacket::read_from(&mut reader).unwra
        });
    });
    c.bench_function("old:medium", |b| {
        b.iter(|| {
            let mut reader = medium_packet.as_slice();
            OldRequestPacket::read_from(&mut reader).unwra
        });
    });
    c.bench_function("old:large", |b| {
        b.iter(|| {
            let mut reader = large_packet.as_slice();
            OldRequestPacket::read_from(&mut reader).unwra
        });
    });
}
struct OldRequestPacket {
    pub header: RequestHeader,
    pub extra: Vec<u8>,
    pub key: Vec<u8>,
    pub value: Vec<u8>,
}
impl OldRequestPacket {
    fn read_from<R: Read>(reader: &mut R) -> std::io::Resu
        let header = RequestHeader::read_from(reader)?;
        let extra_len = header.extra_len as usize;
        let key_len = header.key_len as usize;
        let value_len = header.body_len as usize - extra_l
        let extra = {
            let mut buf = Vec::with_capacity(extra_len as 
            reader.take(extra_len as u64).read_to_end(&mut
            buf
        };
        let key = {
            let mut buf = Vec::with_capacity(key_len as us
            reader.take(key_len as u64).read_to_end(&mut b
            buf
        };
        let value = {
            let mut buf = Vec::with_capacity(value_len as 
            reader.take(value_len as u64).read_to_end(&mut
            buf
        };
        Ok(Self {
            header,
            extra,
            key,
            value,
        })
    }
}
```
Here's results.
```
Benchmarking new:small
Benchmarking new:small: Warming up for 3.0000 s
Benchmarking new:small: Collecting 100 samples in estimated 5.0003 s (32M iterations)
Benchmarking new:small: Analyzing
new:small               time:   [151.18 ns 151.39 ns 151.59 ns]

Benchmarking new:medium
Benchmarking new:medium: Warming up for 3.0000 s
Benchmarking new:medium: Collecting 100 samples in estimated 5.0005 s (30M iterations)
Benchmarking new:medium: Analyzing
new:medium              time:   [161.13 ns 161.52 ns 161.92 ns]
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild

Benchmarking new:large
Benchmarking new:large: Warming up for 3.0000 s
Benchmarking new:large: Collecting 100 samples in estimated 5.0001 s (22M iterations)
Benchmarking new:large: Analyzing
new:large               time:   [221.59 ns 221.95 ns 222.38 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

Benchmarking old:small
Benchmarking old:small: Warming up for 3.0000 s
Benchmarking old:small: Collecting 100 samples in estimated 5.0005 s (25M iterations)
Benchmarking old:small: Analyzing
old:small               time:   [201.33 ns 202.17 ns 203.33 ns]
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe

Benchmarking old:medium
Benchmarking old:medium: Warming up for 3.0000 s
Benchmarking old:medium: Collecting 100 samples in estimated 5.0004 s (24M iterations)
Benchmarking old:medium: Analyzing
old:medium              time:   [207.17 ns 207.55 ns 207.97 ns]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

Benchmarking old:large
Benchmarking old:large: Warming up for 3.0000 s
Benchmarking old:large: Collecting 100 samples in estimated 5.0006 s (20M iterations)
Benchmarking old:large: Analyzing
old:large               time:   [252.29 ns 252.99 ns 253.91 ns]
Found 5 outliers among 100 measurements (5.00%)
  1 (1.00%) high mild
  4 (4.00%) high severe


Process finished with exit code 0
```